refactor: split note_detail.rs into notes/ module (SRP)
