+++ src/ui/notes/mod.rs
@@ -1,0 +1,6 @@
+mod attachments;
+mod detail;
+mod menu;
+mod tags;
+
+pub use detail::NoteDetail;
